requirements.txt
amqp==2.2.1
billiard==3.5.0.3
cachetools==2.0.0
celery==4.1.0
certifi==2017.4.17
chardet==3.0.4
click==6.7
Flask==0.12.2
Flask-SQLAlchemy==2.2
Flask-WTF==0.14.2
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-bigquery==0.25.0
google-cloud-core==0.25.0
googleapis-common-protos==1.5.2
httplib2==0.10.3
idna==2.5
itsdangerous==0.24
Jinja2==2.9.6
kombu==4.1.0
MarkupSafe==1.0
protobuf==3.3.0
pyasn1==0.2.3
pyasn1-modules==0.0.9
pytz==2017.2
redis==2.10.5
requests==2.18.2
rsa==3.4.2
six==1.10.0
SQLAlchemy==1.1.12
urllib3==1.22
vine==1.1.4
Werkzeug==0.12.2
WTForms==2.1
